backend/requirements.txt
beanie~=2.0.1
pydantic~=2.12.4
python-dotenv~=1.2.1
fastapi~=0.123.4
pymongo~=4.15.4
uvicorn~=0.38.0
websockets~=15.0.1
fastapi-crons~=2.0.1